pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["haydendaly <hcd36@cornell.edu>"]

[tool.poetry.scripts]
lint = "src.scripts:lint"
cgc = "src.scrape.cgc:main"
psa = "src.scrape.psa:main"
notebook = "src.scripts:notebook"
setup = "src.scripts:setup"

[tool.poetry.dependencies]
python = "^3.9"
anyio = "3.6.2"
appnope = "0.1.3"
argon2-cffi = "21.3.0"
argon2-cffi-bindings = "21.2.0"
async-timeout = "4.0.2"
attrs = "22.2.0"
backcall = "0.2.0"
beautifulsoup4 = "4.11.2"
cachetools = "5.3.0"
clip = {git = "https://github.com/openai/CLIP.git", rev = "a9b1bf5920416aaeaec965c25dd9e8f98c864f16"}
dataclasses = "0.6"
debugpy = "1.6.6"
decorator = "5.1.1"
huggingface-hub = "0.12.1"
idna = "3.4"
importlib-metadata = "6.0.0"
importlib-resources = "5.12.0"
isoduration = "20.11.0"
jedi = "0.18.2"
Jinja2 = "3.1.2"
joblib = "1.2.0"
jsonpointer = "2.3"
jsonschema = "4.17.3"
lxml = "4.9.2"
Markdown = "3.4.1"
MarkupSafe = "2.1.2"
matplotlib = "3.7.0"
matplotlib-inline = "0.1.6"
mistune = "2.0.5"
nest-asyncio = "1.5.6"
numpy = "1.24.2"
opencv-python = "4.7.0.68"
outcome = "1.2.0"
packaging = "23.0"
pandas = "1.5.3"
pandocfilters = "1.5.0"
pickleshare = "0.7.5"
black = {extras = ["jupyter"], version = "^23.3.0"}
Pillow = "9.4.0"
aiohttp = "^3.8.4"
scikit-learn = "^1.2.2"
torch = "^2.0.0"
torchvision = "^0.15.1"
regex = "^2023.3.23"
python-dotenv = "^1.0.0"
ftfy = "^6.1.1"

[tool.poetry.dev-dependencies]
pytest = "^7.3.0"
pre-commit = "^3.2.2"
isort = "^5.12.0"
notebook = "6.5.2"
notebook-shim = "0.2.2"
nbclassic = "0.5.2"
nbclient = "0.7.2"
nbconvert = "7.2.9"
nbformat = "5.7.3"
jupyter = "1.0.0"
jupyter-console = "6.6.0"
jupyter-events = "0.6.3"
jupyter-client = "8.0.3"
jupyter-core = "5.2.0"
jupyter-server = "2.3.0"
jupyter-server-terminals = "0.4.4"
jupyterlab-pygments = "0.2.2"
jupyterlab-widgets = "3.0.5"
ipykernel = "6.21.2"
ipython = "8.10.0"
ipython_genutils = "0.2.0"
ipywidgets = "8.0.4"

[tool.poetry.group.dev.dependencies]
autoflake = "^2.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
